Adding support for a RPI4B working example
When trying to use the working example for the CM4 in the RPI4B,
the rpi-firmware hangs because it cannot find the RPI4 devicetree.
The rpi-firmware searches for the devicetree filename based on the
board hardware version. Once the vfat partition only contains
the CM4 device tree `bcm2711-rpi-cm4.dts`, the rpi-firmware
cannot find the devicetree (which for the rpi4b is named
bcm2711-rpi-4-b.dts).
This commit aims to add support in the project for the RPI4, so
that the working example based on the CM4 could also be used for
the RPI4.

Signed-off-by: Leonardo Amorim <lam@vizrt.com>